Updated Map Detection
I added the ability for it to auto format map names, and get what shows up in the scoreboard in game. like de_inferno -> Inferno or de_overpass -> Overpass.  No longer need to add new map names to be formated.  Perfect for new maps that get added to the game.